+++ drivers/clk/at91/sckc.c
@@ -315,7 +315,7 @@ static int clk_sam9x5_slow_set_parent(struct clk_hw *hw, u8 index)
 	return 0;
 }
 
-static u8 clk_sam9x5_slow_get_parent(struct clk_hw *hw)
+static int clk_sam9x5_slow_get_parent(struct clk_hw *hw)
 {
 	struct clk_sam9x5_slow *slowck = to_clk_sam9x5_slow(hw);
