+++ drivers/clk/qcom/clk-regmap-mux.c
@@ -23,7 +23,7 @@ static inline struct clk_regmap_mux *to_clk_regmap_mux(struct clk_hw *hw)
 	return container_of(to_clk_regmap(hw), struct clk_regmap_mux, clkr);
 }
 
-static u8 mux_get_parent(struct clk_hw *hw)
+static int mux_get_parent(struct clk_hw *hw)
 {
 	struct clk_regmap_mux *mux = to_clk_regmap_mux(hw);
 	struct clk_regmap *clkr = to_clk_regmap(hw);
